real time update and fontawesome fix
Added real time update of the filters and colors
Added a condition to verify when to request data from the server to
update the genome properties data. So, it will only asks if you
successfuly upload a micromeda file, i.e., if you are connected to the
server. Else, it will do nothing.